Only calculate the original unit cell
We do not need to calculate the LJ potential for atoms outside of the
original unit cell, since they'll just be duplicate calculations.

Making this single change on a box of 2160 framework atoms (in 3x3x2
unit cells) reduces the ipython runtime from 33m23s down to 1m39s,
even allowing us to use a grid spacing of 1.0 Angstrom or finer.  The
metric value is slightly different (by 0.01), but that's normal due
to a different discretization.