Fix histogram bug assuming min(e_val) < 0
Some MOFs (e.g. h5059317) do not have an attractive region (catenation, tight
pores).  This caused a bug in the histogram code: `ValueError: bins must
increase monotonically.`  This commit no longer makes that assumption and
instead touches a `no_histogram.txt` file in cases without attractive regions.
The histogram analysis is also moved to the end of the script in case of
future bugs, so that the raw energy values will be written first.