Aggregate histogram summary statistics
Start looking at the Energy_Values.txt logs to establish reasonable
histogram parameters, so that a first pass on histograms can be
calculated for preliminary machine learning studies.

TODO: encapsulate the get_energy_stats.R code into a function for
more general usage, then modify it to yield histograms within R.
Based on the computation time so far, it seems to be on the order of
minutes to run through the entire 10k structures, so we'll just do the
histograms in post-processing (and export a single file, which will be
easier to work with).