Interactively plot the LJ metric
Use the `manipulate` package in R to interactively adjust the upper and
lower bounds of the LJ metric to look for trends vs. the volumetric
deliverable capacity (g.L).  To me, it looks like there are extremes
that represent Wilmer's vf plots as well as the dragons from the LJ
metric paper.  This will be a useful tool as we try to figure out the
ideal cutoffs for our paper.