Set up grid search for GCMC required
Calculate the number of required simulations to capture all of the top
MOF candidates (e.g. MOFs with a deliverable capacity > 47 g/L). Use a
closure to rewrite the function. Also enclose histogram tabulation
inside helper functions to run them on-demand instead of every time
that the script is opened.