Standardize histogram output
Align histogram bins to the bin_width (e.g. if bin_width is 10, ensure
that all bins are on the tens) and export histograms even for
non-attractive MOFs to make sure that the workflow can still account
for them (especially if we end up writing numerical integrals).

Also fix a regression with the calculation of the reported metric.