Revisit ToBaCCo parity plots with updated and finer grids
Used complete ToBaCCo runs (without missing grid points) and got
slightly improved results, though still not close to the accuracy
within the hMOF database.  Also tried improving the fit with finer
grid spacing, but the predicted output only varied by < 3 g/L, which
doesn't account for the magnitude of error we're seeing in the points
with > 10 g/L of error.