Aggregate energies below the binspec
Added a bin_below option for -Inf, like bin_above for +Inf, which
lumps together all energies below/above the individual bins selected.
This has the advantage of using sufficient data for fitting (the
strongest attractions will be sparsely populated), though more work is
necessary to ensure (1) that the bottom bin is set appropriately, and
(2) that the highly simplified model is reported with the correct
parameters.