Fix implementation of upward/downward for hist spec
Had found an error in the notebook's `coarse_model` code block due to
a missing part of the energy histogram (bins did not sum to one).
After some digging, I found that the histogram binning was still
having the same error as the "hypertuning plot," where there was a
gap between the specified end of the histogram and the actual end
where `bin_below` would pick up.  This has now been corrected.