Start playing with the mixed model
Combining the hMOF/ToBaCCo training sets don't solve all of the
problems, but at least it's easy to describe.  It might be necessary
to go back to one of the offset bin widths for the model.

Also added the textural analyses back into the notebook since some of
the variables were called deeper in the notebook.